Added support for various forms of charts.
This adds pie, bar, stacked, line and area charts.  Simply add
'chart=cname' (where cname is one of the listed types) to the SLIDE
declaration and then put a table into the slide content.  Charts are
generated using Highcharts and are SVG.  Best place for documentation is
section 'four' in the example presentation!

Also I've added support for the kramdown gem which provides the Markdown
table syntax.