Fixing print CSS
When printing the CSS file paths need to be rewritten
to match the absolute URIs used in PDFKit. In addition, it
is required to configure the print media style or 
the output will be buggy. 

In print mode, the position of the slide needs to be 
set explicitly to "relavtive" or more explict styling 
in the slide container will fail in print mode.