presenter view: give the spawned slides window the ability to move our location
This means that you can open the presenter view and have it spawn a
slides window, but then make the slides window fullscreen instead of
the presenter view (which can be a background window on another
display). I didn't test it very much, but it seems to work well enough.

This uses the horrible hack of setting an attribute on the slide
window because otherwise the reference back to the parent window was
sometimes (but not always) getting clobbered when showoff.js loaded in
the child window.