working vagrant-based ubuntu setup
This is the first setup I could reliably get working to allow us
to generate PDFs via Gtk3::WebKit perl, which seems to be The Way
to do interesting and useful things for showoff -> pdf.

This uses vagrant to build an ubuntu vm (running under VirtualBox),
configures the VM with all the necessary packages, and a few CPAN
builds for the tricky parts.

See the README for how to run things.