Handling longer "special content"
Now it's possible to have notes that not only span multiple lines, but
also include blank lines and indented text. For these things (notably
the blank lines), it's necessary to repeat the marker for every line,
and that marker is then taken out.

Moved the content alteration from operating on the final (entire file)
to doing it on a slide-by-slide basis. It was important for the way I
was handling the notes, but more than that I think it's proper to treat
each slide on its own, even if many of them happen to be defined in a
single file.

Also making sure there's no problem if a slide has no notes (either
because the line doesn't exist or there is no note text).