Adding formatting to "special content"
Changing the plain p.notes to div.notes with markdown-formatted content.

Needed to make quite a few changes to the test, mostly in the way of
convenience or helper methods for getting or normalizing the notes. One
of my least-favorite things is trying to compare two bits of HTML to
see if they're equivalent. Even with Nokogiri, this isn't so much fun.

Caught a repeated test name, which was keeping a test from being run
(redefining the test instead of having both defined).