JS/CSS changes for formatted notes
Now that the notes are HTML and not text, the JS that puts the notes in
the right place (for viewers or the presenter) needs to be changed.
Still using just the text for the slide footer the viewers see, but the
presenter now gets the entire notes HTML.

The presenter notes also need to be expanded a bit, allow scrolling for
longer notes, and show lists correctly.