Create showoff->pdf perl script
This is a modification of the s5pdf.pl script from:

  https://github.com/potyl/Webkit/blob/master/s5pdf

which generates PDFs from S5 presentations.  Showoff behaves similarly to S5 and
so with some slight modifications we have something that can be run like:

  xvfb-run --server-args="-screen 0 1200x1024x24" perl showoffpdf.pl

or, if you are running under an X11 display, just:

  perl showoffpdf.pl

This will need more tooling to make it easy to use for our purposes.