A mass of style-choice-related JS
The style menu is modeled on the table of contents, but more complicated
in some cases (getting the menu choices from the links) and simpler in
others (the menu will always be a single level, no nesting).

There's also what's to be done when a choice is selected, which is
"deactivate" (make alternate) all the stylesheet links except for the
chosen one, which is "activated" (made to be just a regular
'stylesheet').