Tying the presenter/regular view styles together.
Now you can get the style menu on the presenter view as well, and a
change to the style on either view (presenter or normal) will be
communicated to the other. The views should now always stay in sync.

The 'prop' argument is for 'propagate', to let the views communicate
back and forth the right amount, without turning this into a massive
stack-busting call loop.