Fixed presenter view to work with popup blocker
Previously, if the popup was blocked, the presenter view just
didn't work at all. Arrow keys did nada.

With this fix, both arrow keys work and a link is presented
to open up a slave window so you don't need to rely on the popup
behaviour. This also lets you open a new slave if you accidentally
close the first.