Add /stats page
This presents a view of the current stats at /stats

The instructor can view a breakdown of who (by hostname) is looking
spending time looking at which pages. Everyone else just gets the
sum of all students.

This does not read in time from last presentation, only from the
current, since you invoked showoff serve for the last time. You should
still yank viewstats.json and save it for later processing.

Please note that the instructor's viewing is not recorded, so you will
see a mostly empty page if you load it up without having viewed the
slides with a computer that is not the instructor's.