Initial support for .instructor tag
This allows the use of the instructor tag.

This will only display the contents of that tag when localhost prints
the /onepage handler. For the time being, it simply suppresses the
output completely. My next revision, when I get time for it, will be to
add new endpoints so that we can print both an instructor guide and a
solutions manual. This will allow authors to use the tag without
buggering up their output though.

Expect the new endpoints in the next few weeks, but I need to get
Extending and the LMS up and running first.