Updated follow mode. Got rid of the localhost crap
New strategy for follow mode. Don't check for localhost anymore.

Now, the presenter window is followed, unconditionally. Other windows
are not tracked. Also updates the ping function and adds a disable
parameter so that windows spawned by the presenter are not tracked.

Next steps are to add a key to "secure" the presenter mode and update
function so that the presenter doesn't have to use presenter mode if
they don't want to.