Added a key parameter
If a "presenter key" is set in your `showoff.json` file, then access to
the /presenter endpoint and ability to update the current page counter
will be limited to browser windows with the ?key=<key> parameter set.

To enable follow mode, append ?key=<key> to the URL in your browser.
This will work in either presenter or standard mode. If there is a
"presenter key" set in `showoff.json` then it must match what you
provide.

I am not happy with this solution, so it is likely to change in the near
future, but for now, this is it.