Added stats & download as popup menus
The stats & download views are loaded dynamically, so they'll be
reloaded each time you hit them. Click the button to open, click again
to close.

It gracefully degrades in the wonky condition in which js is turned off,
even if the rest of showoff is über reliant on js.

You can right click on the button and open in new window to actually get
a new window/tab. You can right click on the files listed to get a link
to the file instead of downloading it. Etc.