Default to binding to all interfaces
Newer versions of sinatra default to binding to localhost only.
This extends a pull request from raphink to bind to the passed in host
value, and also defaults to binding to 0.0.0.0, so anyone can access the
presentation.

If you do want to disable external access, simply pass in -h localhost.