This is a stupid bloody hack of a workaround
I am embarrassed and am less of a programmer for committing this.
However, I am on a plane right now without access to the documentation
to figure out what the hell is going on, and I don't have the time to
design a proper data flow for showoff.

Summary of the problem:

If the `update_special_content` method processes a fragment that
contains an <img /> tag, it will convert it to prehistoric HTML, ie,
without the closing slash. When the HTML parser reparses that fragment
later on, that tag drops it into compatibility mode and HTML
entities are simply discarded.

This means that if you include an image on a slide, then put a code
block after it, the hashrocket (=>) is blown away. Gross.

The real solution is to quit parsing and reparsing the text, but that
will take some rearchitecting and that's not on the roadmap for this
month.

Maybe next.