Generate supplemental materials
Fixes #34

This is a somewhat haphazard fix, in my opinion, but doing it in a
better way requires refactoring request routing. That needs to be done
soon, very very soon, but that day is not today.

To use this, simply add a slide with the classes of 'supplemental' and
whatever type of supplemental material it is.

For example:
  <!SLIDE supplemental solutions>

This can be included in the same file as the labs, so this means that we
can make each lab, its corresponding guide page, and its solution all
self contained.

To access the supplemental materials, simply browse to:
  http://<host>:9090/supplemental/<type>

For example:
  http://localhost:9090/supplemental/solutions