Added file inclusion tag
Include a tag like `~~~FILE:fundamentals-4.1-solution.pp:Puppet~~~` and
it will be replaced with the contents of that file, wrapped in the
proper HTML to display it as code, syntax highlighted with the classes
provided. The general case is these two forms:
* `~~~FILE:<filename>:<space separated list of classes>~~~`
* `~~~FILE:<filename~~~`

Please note that this currently has a bug in which the tag replacement
happens AFTER the markdown parsing. It should happen first, but that
alters the DOM in a strange way that I need to debug first.