alternative to double markdown parsing
Fixes #24

This adds section tags that will work much better than the `.handouts`
and `.notes` tags that we used before. You simply open a section with
`~~~SECTION:<name>~~~` and close it with `~~~ENDSECTION~~~`.

No more need to prepend `.handouts` to each line, and that's where the
benefit comes. Because we're not munging lines, we don't need to parse
markdown twice, so we don't have the crappy double/triple escape bugs.
It also means that we can finally include code in handout notes without
wacky unicode bullshit.

This is going to take a bit of a transitional period, so I'm leaving
both markup algorithms in place for approximately one month while we
start transitioning the material to the new method. When we release the
next revision, I'll push out another update which removes the old
method, which should also speed things up quite a bit.