Add button for reporting issues
If the key "issues" is present in showoff.json, then this will present a
button allowing the presenter to file a ticket from the presenter view.

The issues key should be set to a string that points to the ticketing
system. It will be appended with a URI encoded string indicating which
slide the problem is on. This string will then be opened as a URI in a
new window.