Add audience interaction. Bumps version to 0.9.6
This uses the websocket messaging from the last update to pass other
messages around. This allows the users to:

* Tell the presenter to slow down or speed up.
* Ask the presenter a question
* Provide feedback to the content authors about specific slides.

To access the interface, the user should simply move their mouse near
the left hand side of the screen. The buttons will be disabled and a
disconnected icon will appear in the lower right corner in case of
connectivity problems.

Slide feedback is saved in the feedback.json file, indexed by slide
name. We should determine policy for getting the file back to us.