Rework the slave window
The slave window no longer opens by default. It requires an explicit
click on the `Enable Slave Window` button, which will then pop up the
slave. When you disable it, that window/tab will be closed for you.

This is necessitated by Chrome's habit of disallowing configuration of
popup windows that are not initiated by user action. Sorry.

I hope that it also helps with the referenced bug, though I cannot tell
as I cannot reproduce it for testing.

Possibly addresses #119