Add a code review workflow
If `showoff.json` is configured with an `edit` key pointing to the
repository path, and showoff is executed with the `--review` command
line flag, then a button will be displayed in both the Presenter view
and in the sidebar of the Presentation window allowing team members with
access permissions to directly edit slides using the GitHub source
editor.

Note that if the user has commit privileges, changes will be committed
automatically, and if not, they will be submitted as pull requests.

My intent is to create a `qa/review` branch for these edits to take
place on, so we have a single merge commit to point to in the history.