add named presentation anchors
This means that you can create links in your presentation like
[Go to the Refresher Lesson](#Refresher/Section) and they will
work.

(side effect, you can now link directly to a named slide anywhere)

This is implemented fairly horribly because I'm not a js guru and I'm on
an airplane with no references to figure out the nice clean way. But it
works and it's not too terrible.