This adds the capability for personal notes
In some use cases, it's nice to have the ability to add personal notes
to be displayed alongside the built in presenter notes. This commit
allows you to do this. Just create a short markdown file in `_notes` that
matches the complete pathname of the slide that you'd like to add notes to.

You may need to empty your browser cache before this displays properly
for the first time.