Changing out shjs for highlight.js
shsjs has been unmaintained since 2008. Highlight.js is actively
maintained and has a much much larger library of supported languages.
This commit includes an uncompressed build of *all* supported languages.

The final release will likely have a smaller build, with selected
languages and minified.

This is incomplete. The highlighting fails. But I need to shelf it for
now and move on to another project.