Eliminate the remote shell exploit
Instead of just stuffing code over the wire and blindly executing it,
just send the slide name and the index of the code block on the slide.
The Showoff server will parse the file and load the code to execute
directly from the presentation.

This also adds a shell code evaluator.

Instead of checking for an environment variable, use an actual command
line argument. Pass `-x` or `--execute` when starting Showoff to enable
code execution.