Load the keymap with a relative path so static works
The keymap json was loaded with an absolute path, meaning that when the
static page was loaded from disk, it was trying to read from the
absolute root of the filesystem. Oops :)

This doesn't correct for the `/control` issue. That is scheduled to be
addressed by allowing the interactive features to be disabled
programatically.

Fixes #333