Allow form author to indicate correct option(s)
This extend's the form support to indicate correct options. These will
be highlighted in the presenter view and in the results, both in the
notes and when it's pushed to the display view.

For radio/select, just mark the correct option with an `=` sign, and for
select widgets, surround the correct response with square brackets.

Tangentially, this also slightly restyles the bar graph to make the the
options stand out better.

Updating the parsing engine also fixed another issue, so (yay)

Fixes #332
Fixes #214