Add SSL support and docs
Sinatra & Rack only reasonably support listening on a single port at a
time for applications like this. That means that this will run in either
HTTP mode or in HTTPS mode, but not both.

Simply run with the `--ssl` command line option, or specify it in the
`showoff.json` file.

New settings, both command line and in `showoff.json`:

* `ssl`
* `ssl_certificate`
* `ssl_private_key`