Handle .break and .comment tags in the renderer
Markdown is kind of horrible and gets confused if a code block follows a
list. It needs a block level element between. This handles that by
recognizing a `.break` tag and just deleting it from the output.

Markdown builds all the things appropriately and the block level <p>
convinces it to start the new element, and then we just discard it.

This also lets us do the same thing with `.comment` tags.

Summary: any lines beginning with `.break` or `.comment` will not appear
in the output and will separate lists followed by code.

TODO: all this text munging makes me ill.