Update to highlight.js 9.2.0
This is a major style update. It may break custom styling, so I'ld like
to get it into the v0.12.0 release. This adds a pile of new languages.

One item of interest is that they actually changed the theme names. I've
symlinked the old names, so any presentations that use them should
continue to work properly.

Fixes #402