Remove the /pdf endpoint
This ports the `/pdf` endpoint and ports that functionality to a new
`showoff pdf` command-line option. If users want the old functionality
of providing a PDF for download, they can save the generated PDF into
the `_files/share` directory. This allows the PDF to be generated once
instead of once for every user who wants it.

For example, `showoff pdf _files/share/mypreso.pdf`