Make the last TOC entry discoverable
When the last item is clicked, before this commit it was very hard to
see it open. Now, it will scroll the children menu entries into view.
This is a little weird on the display view, because the menu jumps a
little on that last entry, but there wasn't really a good way to make it
happen only in the presenter view.

YAY for inline functions! @marrero984 should be ashamed! haha

Fixes #385