Allows you to choose the section type to use
This adds two related features. Both of the features work with section
tags, such as:

    ~~~SECTION:notes~~~
    This text is inside a "notes" section.
    ~~~ENDSECTION~~~

Prior to this, there were two types of sections, `notes` and `handouts`.
Now, the types are arbitrary. You can add any section types you want by
simply declaring them in a slide.

When viewing the presenter, any time there are more than one sections
attached to a slide, you'll now be presented with a flyaway tab widget
that will allow you to choose any of the sections to display in the
`#notes` field. The choice is persistent, so if you switch to the
`guide` section, then it will stay that way until you change it.

When printing, you can now choose the section type to attach as notes.

    # Prints the presenter notes
    showoff static print notes

    # Prints a section called "guide"
    showoff static print guide

The existing section types, `notes` and `handouts`, are retained in that
the default type to show in the `#notes` field is the `notes` section
and the default type to print is the `handouts` section. Personal notes
will also only attach to the `notes` section type.