Add dockerdockerdocker stuff
This adds rake tasks to build and push docker images so we can easily
keep them maintained. This uses my namespace for now, since it's
unlikely that we'll get access to the `puppetlabs` namespace on hub.

Maybe we should register `puppetlabs-education`? Dunno, this seems ok
for now.

This also provides an example `Dockerfile` for people to build their own
images containing their presentations.