Reload the page when content is reloaded
This will refresh styles along with content updates when you trigger the
refresh function. It adds a negligible amount of time to the reload,
since most of that was in presentation compilation anyway. It also
provides feedback that *something* is happening, instead of just a long
pause and then slides redrawing.

Fixes #534