Made reloading content more reliable
We recently added full reloading of the UI (#535) so that styles, etc.
were updated along with content when the `reload` function was triggered.
Unfortunately, that had the side effect of redrawing content before the
new presentation had completed compilation. In other words, it would
reload with the old content, and you'd have to wait a few seconds and
hit it again before you'd see changes.

This corrects that. It also cleans up some boogered indentation.

Fixes #540