Adds primitive annotation support
Kari was away for the weekend, so I decided to hack up something fun.
This is not quite ready for merge, but it's rather usable in its current
state. It allows the instructor to draw on the slide from the presenter
view. The annotations are replicated on the display view on the
projector and each of the audience member's browsers. Annotations are
attached to the slide rather than just overlaying the whole display.
This means that switching slides will switch annotations too.

* It's untested on IE.
* It boogers up the section pages.
* The scaling between different views is a tiny bit off. (damn you, zoom!)
* It could use a little minor polishing.