Add simple CSS property detection
This will redirect to an error page on super old browsers and will add a
'no-zoom' class to the body tag on Firefox. At some point, we might
consider Modernizer, but I think that's overkill for what we need right
now.

This will support moderately cleaner layout workarounds for funky
browsers. Such as all of them.