Correct the out-of-control preview size on firefox
With `#preso` positioned normally, it affected the size of its parent.
that meant that the container was pushed to a size to fit around the
slide, then we tried to use that size to scale the slide down to fit.
Since it was in a container sized for it, it already fit. But everything
else was boogered.

Setting the `#preso` element to absolute positioning, this pulled it out
of the normal flow, so it didn't affect the layout size of its parent.
That means that our scaling algorithm now works properly.

What a pain in the ass. I will be happy when we figure out a better way.