big-ified text on slides
Tag a slide with `bigtext` and its context will grow to the maximum size
possible. This will only look decent on slides with just a few words or
short phrases! It looks best when combined with `center`.

I don't really like big-ifying on each slide display, but it appears
that the size is calculated via visibility, so we kind of have to re-run
it each time a new slide becomes visible.

There may be a better way of doing this.

Fixes #608