Change the checkboxes to toggle buttons
I'm not happy with the fiddling I had to do to position these buttons
right. It seems fragile to me. But it looks right on Safari/Chrome/Firefox.
It seems rather heavy for such a small bit of styling, but I trust that
we'll start using JQuery UI more heavily in the future and can start
replacing some of the hand-rolled widgets.

Fixes #623