Highlight fenced code blocks properly
When fenced code blocks are used, eg with Commonmarker, the Showoff
language tags weren't used. This means that the logic that adds the
proper classes weren't added. This refactors that check so that the
classes are added to the `<pre>` tag either way, & highlighting now
works again.

Fixes PR #646 in a less disruptive manner.