Makes the viewing stats more useful
The datadump of the old viewstats menu didn't help anyone. Instead, this
shows an interactive sparkline graph of what the audience is currently
looking at and another graph of elapsed time for all slides.

This makes it trivial to visualize how you're capturing their attention
and makes it clear when you should revisit some topics.

I didn't spend a whole ton of time on graphical styling, since you're in
the middle of the restyle. Also, it looks a lot more interesting when there
are more than one or two clients!

Features:

* Highlights the menu button when more than 25% of the audience isn't
  following along.
* Click on the bars to navigate the presenter to that slide.
* Hover a bar to see its title and how many people are on it, or the
  total elapsed time on it.
* Open in new page to see full details of the elapsed time. This is the
  old "all" view. I didn't really do anything to it because I don't know
  that I think it's worth keeping around.
* Menu live updates every 30 seconds.
* Graphs zoom the bar on hover when it's too small to click. (it's still
  kind of a pain on the elapsed time graph, so it could use some love.)
* Graph is plain html, fully stylable and works even on IE.