Switches internal client indexing to GUID
Prior to this it was using IP address, which made it unreliable when
clients were behind a proxy or the like. This also makes it easier to
test, since different browsers on a developer's machine will show up as
different connections.

This is visible to the presenter in the stats presentation and the forms
answers.

Fixes #539