Add an idle indicator
This will put a notice if more than 25% of the audience hasn't looked at
a slide in more than 5 minutes. This will be 100% of the audience during
longer slides, for example lab slides in the PL courseware in which the
slide may not change for 15 or 20 minutes.

This stat will not trigger the warning state on the toolbar button.