WIP: Updated section format
This allows the sections to be specified by name instead of just by the
directory the slide happens to live in. This lets you build a
presentation structure that doesn't necessarily map to your directory
structure.

This will load all existing formats. We can deprecate old format later.

This *should* be complete, but I'd like to do more testing before
releasing it to the wild.

Fixes #678