Getting started with internationalization
This integrates the Ruby standard i18n library. All obvious strings in
the web UI generated via either ERB templates or Ruby code are translated
directly with direct I18n calls. All strings are exposed as a giant hash
to Javascript and a small helper method I wrote to give it the same
calling syntax as in Ruby. This seemed more expedient than integrating
another different library for the JS side of things.

This does not yet attempt to internationalize the CLI strings. I am not
fully convinced of the ROI on that yet.