Add the ability to localize the content
Just translate the content and place it in `locales/[lang code]`. For
example, `locales/es` or `locales/ja`. The slide names and relative
structure must be exactly the same, because it is using the same
`showoff.json` presentation description. Any images should also be
located in that directory alongside the content. This allows you to
localize images of text as well. If you don't need that, you can just
symlink it to the toplevel `_images` directory.