Unified custom syntax templates
Moves most of the custom syntax handlers into a separate method. This
should make things more maintainable.

This handles the custom elements by adding html tags to the markdown.
It adds newlines around the html tags which enables commonmark to parse
the markdown between the tags and the attribute markdown="1" to allow
kramdown to do the same thing.