content language selector for the presenter
This allows the presenter to select a content translation, when one is
available. It defaults to "auto" which means use the browser setting and
offers the ability to disable translation, or just present the original
language.

Next up is to add the widget to the audience view sidebar and more
testing.