Track the "master" presenter
The first browser that loads the presenter view is recorded as the
"master". Any other browsers that load it will default to not updating
the follower. This is useful, for example, if you have a second
presenter observing the presentation.

You can still enable this from the secondary presenters, it just
defaults to off.