Add content language chooser
If the presentation has multiple translations, this will add a dropdown
in the settings modal for the presenter and the sidebar for the audience
view. The choices are all translations, auto (let the browser choose),
and disabled (always use the original translation).

The UI language is not configurable in this fashion. It always follows
the browser configuration.