Add the ability to translate simple strings in SVG
See the docs for more info on how to use it, but this just adds the
ability to do simple text swapping translations on SVG images, either
included directly inline or as img tags (which are inlined so we can
manipulate them via the DOM). Since the mechanism for doing so was so
similar, I also added the ability to add translated string tags.