Add button to sync with the presenter
When an audience member navigates to a slide that is not the same as the
one being presented, a button will appear providing a quick shortcut to
jump back to catch up with the presentation. This also appears if the
audience member fails to mark an activity slide complete before the
presenter navigates away from it.