Explicitly pin the page to the TOP of the viewport
Chrome and Firefox on Linux pin this div to the BOTTOM of the viewport
when it's too big. This means that the top of the page scrolls out of
view. Which sucks, since that's where the controls are.

All other browsers, including the same browsers on other platfomr, seem
to pin the div to the BOTTOM, meaning that the bottom scrolls out of
view under the same circumstances.

Fixes #776