Added a first run tour of the presenter view
The first time the presenter is loaded, this will run through a
whirlwind tour of the interface. Once a tour has been completed (all
the way to clicking *Done*), then it won't show again.

This has a couple nice added benefits. Each stop in the tour has a
version number associated. When the tour is marked completed, the
version is recorded. That makes it progressive. We can add items later
on and increment the tour version. This would mean that currect users would
only see the *new* items, but new users would see the whole tour.

It also filters out interface elements that don't exist. For example, if
a presentation doesn't configure issue reporting, then that button
doesn't appear and so the tour skips over it.

TODO: I can't do the translations yet because I'm on a plane.