Now displays possible translations when checking
When running `rake lang:check`, it will now do a google translate call
for you when a string key is missing in a translation file. I did *not*
have it auto-update the translation file because that would lose any
formatting we've added to the file and I'd rather a human looked at the
translation to at least somewhat sanity check it.