WIP: add ability to create arbitrary intro tours
This provides a method which can be called to display arbitrary named
tours--but only if it hasn't already been seen. An example usage would
be to show a quick tour of the forms interface for the presenter when
it's first displayed.

TODO: this shouldn't live in the presenter.erb, and this could maybe be
refactored a bit to share more with the original tour implementation.
This cannot be merged until some issues regardign stacking transformed
elements are resolved upstream. (Showing a tour on an in-slide element
causes the slide to zoom out to full size, no matter the screen size.)

https://github.com/usablica/intro.js/issues?utf8=✓&q=is%3Aissue%20is%3Aopen%20transform