Make introjs work on slides.
The upstream library forced elements to `transform: none` in order to
avoid the stacking context problem. (z-index only matters within a
stacking context, and each transformed element created a new stacking
context. So there's no way to get fine grained control of z-index for
elements *on a slide*.)

This corrects that problem by approaching it in a different way. When we
cannot affect the z-index, we instead duplicate the element we/re
highlighting and put that in front. The cloned element is discarded later.

This local change is captured upstream by https://github.com/usablica/intro.js/pull/727