Finishes the ability to add arbitrary tours
The presenter and the audience will get information appropriate to their
roles about as new UI elements become visible for the first time.

This also adds the ability for presentation authors to include tours in
their presentations with just a tag and a snipped of javascript, but I'm
not going to expose that in documentation until we've battle tested it a bit.