Don't show tour by default
This reworks the tour code a little bit so that we can show the intro on
demand. Since cookies are set by domain, there's no way to know whether
the presenter has actually seen the tour before unless we have a
consistent domain. So for now, only automatically show the tour if we're
presenting from `localhost`. Instead, presenters get a button in the
help dialog that starts the tour for them.

Audience members still get the tour, since it's not likely that they'll
see many different Showoff presentations from many different servers.

Fixes #816