WIP: interstitial loading screen
This presents a spinner while the presentation is loading, rather than a
skeleton of a presentation that looks like half of it is just gone.

It's marked at WIP because the animation isn't 100% smooth in all cases.
For example, Chrome chunks a bit when the slides are rendering and
fading in. I will likely redo how the opacity is implemented.