Use temp colon separated lang name string for commonmarker
Commonmarker ignores all but the first word in a fenced code block. This
takes all those classes, smooshes them into a colon separated string
for markdown rendering and then explodes them back into classes.

What do you think about this approach for #876, @raphink? I don't know
that I'm happy with the location of the hacks, but until I refactor into
more modular classes this seems to be the best place to do this.