Barebones presentation render
This will take a `showoff.json` and render each slide into the html
content of a presentation. This is just content, not the slide
transitions or the document generation, or any of the special Showoff
markdown extensions. In short, it's not useful yet, it's just a proof of
concept.

Run `showoff --dev static` to see it in underwhelming action. It just
dumps the rendered HTML and then a quick tree of the sections & slides.